Revert "Merge branch 'fix/dev-13414' into ftr/DEV13471"
This reverts commit 34cefbd7e17382c5b045050251f78ee24c1aa9e0, reversing
changes made to a65cc7f552dfcf0e9fa4094cb607973a3c121143.
