+++ src/_scss/components/checkboxTree/_checkboxTree.scss
@@ -277,106 +277,6 @@
   width: 100%;
 }
 
-// New nested checkbox tree
-.psc-checkbox .checkbox-tree {
-  ul {
-    margin: 0;
-    padding: 0;
-
-    li {
-      list-style-type: none;
-
-      .label {
-        display: flex;
-      }
-    }
-
-    .loading {
-      margin-left: 10px;
-    }
-  }
-
-  div.open {
-    margin-left: 4px;
-    padding-left: 24px;
-    border-left: 1.5px solid $gray-cool-20;
-  }
-
-  button {
-    width: unset;
-    background-color: unset;
-    margin: unset;
-    padding: unset;
-    color: $color_gray;
-    margin-right: 4px;
-    svg {
-      width: 10px;
-      height: 10px;
-      color: $blue-50;
-    }
-  }
-
-  input[type="checkbox"] {
-    margin: 2px 4px 0 0;
-    cursor: pointer;
-  }
-
-  .checkbox-tree-label {
-    &__container {
-      display: flex;
-    }
-
-    &__controls {
-      display: flex;
-      align-items: center;
-      margin-top: 4px;
-      height: 20px;
-    }
-
-    &__value-container {
-      display: inline-flex;
-      align-items: center;
-    }
-
-    &__description {
-      margin-left: 10px;
-    }
-
-    &__value-container-value {
-      font-weight: $font-semibold;
-    }
-
-    &__value-container-count {
-      font-style: italic;
-      white-space: nowrap;
-    }
-    &__value-container-value {
-      color: $gray-80;
-
-      span {
-        font-size: $font-size-10;
-        font-style: italic;
-        line-height: rem(1.8);
-        color: $gray-60;
-      }
-      span.highlight {
-        background-color: $yellow-10;
-        font-style: normal;
-        font-size: $font-size-12;
-      }
-    }
-
-    &__container {
-      margin-left: 4px;
-      //padding-left: 24px;
-
-      .open {
-        border-left: 1.5px solid $gray-cool-20;
-      }
-    }
-  }
-}
-
 // searchV2
 @import "./searchV2CheckboxTree";
 
@@ -386,3 +286,5 @@
 // def code bulk account download
 @import "./defCodeBulkDownload";
 
+// tree nodes scss
+@import "treeNodes";
