+++ src/_scss/components/checkboxTree/_treeNodes.scss
@@ -1,0 +1,9 @@
+.checkbox-tree {
+  button {
+    width: unset;
+    background-color: unset;
+    margin: unset;
+    padding: unset;
+    color: $color_gray;
+  }
+}
