+++ src/js/components/sharedComponents/CheckboxTreeLabel.jsx
@@ -1,7 +1,6 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import replaceString from 'helpers/replaceString';
 
 const propTypes = {
     value: PropTypes.oneOfType([
@@ -16,8 +15,7 @@ const propTypes = {
     displayId: PropTypes.bool,
     countLabel: PropTypes.string,
     labelClassName: PropTypes.string,
-    subLabel: PropTypes.string,
-    searchString: PropTypes.string
+    subLabel: PropTypes.string
 };
 
 const CheckboxTreeLabel = ({
@@ -27,37 +25,32 @@ const CheckboxTreeLabel = ({
     displayId = true,
     countLabel = '',
     labelClassName = '',
-    subLabel = null,
-    searchString
+    subLabel = null
 }) => {
     const countText = count > 1 ? 'codes' : 'code';
     const countDisplay = countLabel === '' ? countText : countLabel;
     const className = labelClassName ? ` ${labelClassName}` : '';
-    const highlightText = (text) => replaceString(text, searchString, 'highlight');
-
     return (
         <div className={`checkbox-tree-label${className}`}>
             {displayId && (
                 <div className="checkbox-tree-label__value-container">
                     <div className="checkbox-tree-label__value-container-value">
-                        {highlightText(value)}
+                        {value}
                     </div>
                     <div className="checkbox-tree-label__value-container-count">
                         {count ? `${count} ${countDisplay}` : ''}
                     </div>
                 </div>
             )}
-            {label && (
-                <div className="checkbox-tree-label__label">
-                    {highlightText(label)}
-                    {subLabel && (
-                        <>
-                            <br />
-                            {highlightText(subLabel)}
-                        </>
-                    )}
-                </div>
-            )}
+            <div className="checkbox-tree-label__label">
+                {label}
+                {subLabel && (
+                    <>
+                        <br />
+                        <span>{subLabel}</span>
+                    </>
+                )}
+            </div>
         </div>
     );
 };
