+++ src/js/components/sharedComponents/checkboxTree/CheckboxTree.jsx
@@ -9,7 +9,9 @@ import { difference } from 'lodash-es';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import replaceString from 'helpers/replaceString';
 import CheckboxTreeLabel from 'components/sharedComponents/CheckboxTreeLabel';
-import TreeNodes from 'components/sharedComponents/checkboxTree/TreeNodesWrapper';
+import TreeNodes from 'components/sharedComponents/checkboxTree/TreeNodes';
+
+import 'react-checkbox-tree/lib/react-checkbox-tree.css';
 
 const propTypes = {
     data: PropTypes.array,
@@ -19,7 +21,7 @@ const propTypes = {
     isDisabled: PropTypes.bool,
     errorMessage: PropTypes.string,
     isSearch: PropTypes.bool,
-    searchString: PropTypes.string,
+    searchText: PropTypes.string,
     modifyLabelTextClassname: PropTypes.string,
     labelComponent: PropTypes.element,
     onExpand: PropTypes.func,
@@ -40,7 +42,7 @@ const CheckboxTree = ({
     isDisabled = false,
     errorMessage,
     isSearch,
-    searchString,
+    searchText,
     modifyLabelTextClassname,
     labelComponent,
     onExpand: onExpandProp,
@@ -175,7 +177,7 @@ const CheckboxTree = ({
      * or string if no match is found.
      */
     const highlightText = (text) => replaceString(
-        text, searchString, modifyLabelTextClassname || 'highlight'
+        text, searchText, modifyLabelTextClassname || 'highlight'
     );
 
     /**
@@ -185,7 +187,7 @@ const CheckboxTree = ({
       * @returns {Array.<object>} An array of objects
     **/
     const createLabels = (nodes) => nodes.map((node) => {
-    // if label is not a string, do nothing
+    // if label is a string, do nothing
         if (typeof node.label !== 'string') return node;
         if (node.isPlaceHolder && node.className !== 'hide') {
             return {
@@ -217,8 +219,7 @@ const CheckboxTree = ({
                         label={node?.isSearchable === false
                             ? node.label
                             : highlightText(node.label)}
-                        countLabel={countLabel}
-                        searchString={searchString} />
+                        countLabel={countLabel} />
                 ),
             children: node.children
                 ? createLabels(node.children)
@@ -233,7 +234,7 @@ const CheckboxTree = ({
             <div className="checkbox-tree-filter-message-container">
                 <FontAwesomeIcon icon="spinner" spin />
                 <div className="checkbox-tree-filter-message-container__text">
-                    Loading your data...
+                    Loading your dataLoading your data...
                 </div>
             </div>
         );
